Extract RubyIndexer::TestCase
